preprocessing performs path extraction
